ArwDecoder::SonyDecrypt(): explicitly truncate the 'key'.
This appears to be the correct expected behavior.

../src/librawspeed/decoders/ArwDecoder.cpp:374:20: runtime error: implicit conversion from type 'unsigned long' of value 140849956640625001 (64-bit, unsigned) to type 'rawspeed::uint32' (aka 'unsigned int') changed the value to 271683945 (32-bit, unsigned)
    #0 0x2a7caa in rawspeed::ArwDecoder::SonyDecrypt(unsigned int const*, unsigned int*, unsigned int, unsigned int) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/ArwDecoder.cpp:374:20
    #1 0x2a7626 in rawspeed::ArwDecoder::decodeSRF(rawspeed::TiffIFD const*) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/ArwDecoder.cpp:84:3
    #2 0x2a7ed3 in rawspeed::ArwDecoder::decodeRawInternal() /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/ArwDecoder.cpp:134:14

+++ src/librawspeed/decoders/ArwDecoder.cpp
@@ -371,7 +371,7 @@ void ArwDecoder::SonyDecrypt(const uint32* ibuf, uint32* obuf, uint32 len,
 
   // Initialize the decryption pad from the key
   for (int p=0; p < 4; p++)
-    pad[p] = key = key * 48828125UL + 1UL;
+    pad[p] = key = uint32(key * 48828125UL + 1UL);
   pad[3] = pad[3] << 1 | (pad[0]^pad[2]) >> 31;
   for (int p=4; p < 127; p++)
     pad[p] = (pad[p-4]^pad[p-2]) << 1 | (pad[p-3]^pad[p-1]) >> 31;
